# api/commander-openapi.yaml
openapi: 3.0.4
info:
  title: Commander API
  version: 0.0.1
  description: Control plane API for managing uploaded functions.
servers:
  - url: http://localhost:8080

paths:
  /functions:
    post:
      operationId: createFunction
      summary: Upload a function binary
      description: Creates a function record and uploads the binary payload using multipart form data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Human readable function name.
                binary:
                  type: string
                  format: binary
                  description: Function binary.
              required: [name, binary]
      responses:
        '201':
          description: Function uploaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateFunctionResponse'
        '413':
          $ref: './common.yaml#/components/responses/ErrorDetailResponse'
        '500':
          $ref: './common.yaml#/components/responses/ErrorDetailResponse'

  /functions/{id}/binary:
    get:
      operationId: downloadFunctionBinary
      summary: Download a function binary
      description: Streams the uploaded function binary and returns function metadata in response headers.
      parameters:
        - name: id
          in: path
          required: true
          description: Unique function identifier.
          schema:
            type: string
      responses:
        '200':
          description: Function binary streamed successfully.
          headers:
            Function-Id:
              description: Unique function identifier.
              schema:
                type: integer
                format: int64
            Function-Name:
              description: Human readable function name.
              schema:
                type: string
            Function-Filename:
              description: Function binary filename.
              schema:
                type: string
            Function-Path:
              description: Original function binary path in commander storage.
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: './common.yaml#/components/responses/ErrorDetailResponse'
        '500':
          $ref: './common.yaml#/components/responses/ErrorDetailResponse'

components:
  schemas:
    Function:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique function identifier.
        name:
          type: string
          description: Human readable function name.
      required: [id, name]
    CreateFunctionResponse:
      $ref: '#/components/schemas/Function'